Spike making it possible for users to sign up to multiple or joined subscriptions and hide the complexity
Added comments ton help me understand how it works.
Added OrJoinedSubscriberList which is what the
user will actually be signed up to.
The basic idea is that when an immediate change is
matched to a SubscriberList that is part of an OrJoinedSubscriberList then the
MatchedContentChange will link to the OrJoinedSubscriberList
not the SubscriberList, thus ensuring that only one email
will be sent.
Additionally it means that in order to manage the
subscription the OrJoinedSubscriberList will
manage it's SubscriberLists, thus making things
fairly transparent for other parts of the app