Move courtesy copies for travel advice to the high queue
Emails to travel advice subscribers are enqueud to
delivery_immediate_high. This queue has the highest priority.

Courtesy copies for travel advice are currently enqueued to
delivery_immediate.

If there is lots of publishing activity (for example from whitehall),
the delivery_immediate queue can become quite large. As travel advice
courtesy copy emails are enqueued they become mixed with emails for all
other publishing.

As the delivery_immediate_high queue is processed with a higher priority
than delivery_immediate, we can end up sending travel advice emails to
all subscribers, but still not have sent the courtesy copy email.

When we check the courtesy copy inbox for travel advice and find recent
travel advice missing, we alert and page developers, both in and out of
hours.

The intent of the alert is to warn developers that there might be
something wrong with email-alert-api. However, with the recent high
volume of publishing, we can end up with travel alerts sent to
subscribers having completed while the courtesy copy has not yet been
sent. When delivery_immediate queue is large, there can be a
considerable delay to the travel advice reaching the courtesy copy
inbox. In these cases, we page on-call developers even though there is
very little they can do but wait for queues to clear.

Since travel advice and the travel advice courtesy copy are processed on
different queues with different priorities, slow email
delivery to the courtesy copy group does not represent slow email
delivery to travel advice subscribers.

Instead, enqueue travel alert courtesy copies to delivery_immediate_high.
This places them on the same queue as all other travel alert emails.

This gives a more accurate representation of actual travel alert emails
to actual subscribers and will hopefully quieten some alerts that have
paged people unnecessarily.

I considered, instead, creating another queue specifically for travel
alert courtesy copies that would have a priority between delively_immediate_high
and delivery_immediate, but that seemed like more complexity for little
reward.

I also considered using the priority of the ContentChange to drive the queue
for the courtesy copy (as it does for normal emails[0] - a priority of high
is put on the delivery_immediate_high queue), but I wanted to scope
this only to travel advice for now since it is the one where we are
seeing issues.

[0] https://github.com/alphagov/email-alert-api/blob/e9ff766bd14cbc966ff5a7d94aca19227060c134/app/workers/process_and_generate_emails_worker.rb#L124