Implement OIDC auth flow
This adds two new endpoints:

    GET  /subscribers/oidc  generate an OIDC login URI
    POST /subscribers/oidc  validate an OIDC code

Like the `POST /subscribers/auth-token` endpoint, this allows
authenticating a user and returning their user ID to another app.
This patch requires the following environment variables:

- `OIDC_PROVIDER_URI`  - used for discovery
- `OIDC_CLIENT_ID`     - identifies email-alert-api
- `OIDC_CLIENT_SECRET` - for the encryption

The secret doesn't need to be known by the app which calls these new
endpoints, as all the validation is done here, so we retain the
guarantee that a user has consented before we give out their ID (and
so open access to other API operations, like managing subscriptions).

The destination URI is unconstrained here, but an OIDC identity
provider will generally have an allowlist.  So the OIDC IDP will
essentially maintain a list of things which are allowed to access
email-alert-api.