Break email-alert-api
We want to break email-alert-api in the same fashion as what occured
when we had an incident due to a faulty deploy. Deploying a branch to
integration should give us what indicaters went off telling us there was
a problem.

At the minute this is a bit hazy. We then want to improve our alerting
so it's more obvious something is majorly broken.