Add processed_at column to digest_runs
The purpose of this column is to store the time that a DigestRun has
completed the initial processing. A nil value in this table indicates
that this DigestRun hasn't yet been processed and the worker for this
job should try again as the system is in an inconsistent state.

The rather arbitrary sounding name, processed_at, reflects similar
namings on ContentChange and Message concepts.

This has been added in so that the application can be more resilient.
Currently if an error occurs during the initiation of a DigestRun we
have no evidence that the run was incomplete and people will just not
receive emails.