Restrict deleting subscribers if emails exist
We want to change the behaviour of the db to restrict the deletion of
subscribers if they have associated emails, the previous behaviour was
to cascade the emails.

We don't want to do that anymore as we introduce a job to cleanup old
[subscribers] as emails shouldn't exist at the point of subscriber
deletion and if they do then we should raise an error.

Trello:
https://trello.com/c/tfu9zXuy/534-delete-unused-data-after-a-year

[subscribers]: https://github.com/alphagov/email-alert-api/pull/1432